Rename FileEncryptor file
