+++ src/crypto/file-encryptor.class.ts
